+++ train_fedavg.py
@@ -418,28 +418,28 @@ def main_worker(gpu, ngpus_per_node, args):
             'test acc', acc.cpu().data.numpy(), global_step=r)
         val_writer.add_scalar(
             'best_acc1', best_acc1.cpu().data.numpy(), global_step=r)
+        if(args.save_weight):
+            # save checkpoints
+            filename = "checkpoint_{0}.pth.tar".format(r)
+            saved_ckpt_filenames.append(filename)
+            # remove the oldest file if the number of saved ckpts is greater than args.max_ckpt_nums
+            if len(saved_ckpt_filenames) > args.max_ckpt_nums:
+                os.remove(os.path.join(args.model_dir,
+                                    saved_ckpt_filenames.pop(0)))
+            if(args.save_weight):
+                ckpt_dict = {
+                    'round': r + 1,
+                    'arch': args.arch,
+                    'state_dict': global_model.state_dict(),
+                    'best_acc1': best_acc1,
+                    'optimizer': optimizers[i].state_dict(),
+                }
 
-        # save checkpoints
-        filename = "checkpoint_{0}.pth.tar".format(r)
-        saved_ckpt_filenames.append(filename)
-        # remove the oldest file if the number of saved ckpts is greater than args.max_ckpt_nums
-        if len(saved_ckpt_filenames) > args.max_ckpt_nums:
-            os.remove(os.path.join(args.model_dir,
-                                   saved_ckpt_filenames.pop(0)))
+                if args.is_amp:
+                    ckpt_dict['scaler'] = scaler.state_dict()
 
-        ckpt_dict = {
-            'round': r + 1,
-            'arch': args.arch,
-            'state_dict': global_model.state_dict(),
-            'best_acc1': best_acc1,
-            'optimizer': optimizers[i].state_dict(),
-        }
-
-        if args.is_amp:
-            ckpt_dict['scaler'] = scaler.state_dict()
-
-        metric.save_checkpoint(
-            ckpt_dict, is_best, args.model_dir, filename=filename)
+                metric.save_checkpoint(
+                    ckpt_dict, is_best, args.model_dir, filename=filename)
         print('%d-th round' % r)
         print('average train loss %0.3g | test loss %0.3g | test acc: %0.3f' %
               (loss / args.num_selected, test_loss, acc))
